# Fix routing conflict for dynamic routes and static route with common prefix (#1509)