This PR resolves issue #1509 

A routing conflict is caused by radix tree node splits on common prefix of static routes together with a param route on the same level. An additional test has been added for this issue.

This patch actually ensures that routing will continue back up the route tree in that case.