@lammel 2020-02-19
This PR has an impact on the routing performance of ~15% in the worst case.
Numbers are taken using the available test benchmarks on my local system.

Before the patch (master, f4b5a90a):

```
sh# /usr/bin/go test -benchmem -run=^$ github.com/labstack/echo/v4 -bench ^(BenchmarkRouterStaticRoutes|BenchmarkRouterGitHubAPI|BenchmarkRouterParseAPI|BenchmarkRouterGooglePlusAPI)$

goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkRouterStaticRoutes-16     	   66543	     17475 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPI-16        	   31058	     38630 ns/op	       2 B/op	       0 allocs/op
BenchmarkRouterParseAPI-16         	  174482	      5884 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPI-16    	  123810	      9144 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/labstack/echo/v4	5.274s
```

With this patch applied:

```
sh# /usr/bin/go test -benchmem -run=^$ github.com/labstack/echo/v4 -bench ^(BenchmarkRouterStaticRoutes|BenchmarkRouterGitHubAPI|BenchmarkRouterParseAPI|BenchmarkRouterGooglePlusAPI)$

goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkRouterStaticRoutes-16     	   56984	     20068 ns/op	       1 B/op	       0 allocs/op
BenchmarkRouterGitHubAPI-16        	   26817	     44528 ns/op	       2 B/op	       0 allocs/op
BenchmarkRouterParseAPI-16         	  192571	      5751 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPI-16    	  100843	     11578 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/labstack/echo/v4	5.476s
```