@lammel 2020-02-19
Managed to save some CPU cylces by avoiding backing up the tree for static only route trees (where no next param or any nodes are stored)
```
goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkRouterStaticRoutes-16     	   63564	     18446 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPI-16        	   27559	     43419 ns/op	       2 B/op	       0 allocs/op
BenchmarkRouterParseAPI-16         	  194587	      5777 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPI-16    	  107343	     11036 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/labstack/echo/v4	5.501s
```